+++ example/example.py
@@ -1,19 +1,21 @@
 from pathlib import Path
 
 from birdnet import get_species_from_file
-from birdnet.models import ModelV2M4
+from birdnet.models.v2m4 import AudioModelV2M4
+from birdnet.types import SpeciesPredictions
 
 audio_path = Path("example/soundscape.wav")
 species_path = Path("example/species_list.txt")
 
 # create model instance for v2.4
-model = ModelV2M4(language="en_us")
+model = AudioModelV2M4(language="en_us")
 
 # predict only the species from this file
 custom_species = get_species_from_file(species_path)
 
 # predict species for the whole audio file
-predictions = model.predict_species_within_audio_file(audio_path, filter_species=custom_species)
+predictions = SpeciesPredictions(model.predict_species_within_audio_file(
+    audio_path, filter_species=custom_species))
 
 # get predictions at time interval 0s-3s
 first_chunk_predictions = list(predictions[(0.0, 3.0)].items())