+++ src/birdnet/models/v2m4/model_v2m4_raven_custom.py
@@ -8,8 +8,9 @@ import tensorflow as tf
 from ordered_set import OrderedSet
 
 from birdnet.models.v2m4.model_v2m4_protobuf import (AudioModelV2M4ProtobufBase,
-                                                check_protobuf_model_files_exist, get_custom_device,
-                                                try_get_gpu_otherwise_return_cpu)
+                                                     check_protobuf_model_files_exist,
+                                                     get_custom_device,
+                                                     try_get_gpu_otherwise_return_cpu)
 from birdnet.types import Species
 from birdnet.utils import sigmoid_inverse