+++ src/birdnet_tests/models/m2v4/model_v2m4_protobuf_py/test_predict_species_within_audio_file_cpu.py
@@ -42,10 +42,10 @@ def test_identical_predictions_return_same_result(model: AudioModelV2M4Protobuf)
 def test_invalid_species_filter_raises_value_error(model: AudioModelV2M4Protobuf):
   invalid_filter_species: Set[Species] = {"species"}
   with pytest.raises(ValueError, match=rf"At least one species defined in 'filter_species' is invalid! They need to be known species, e.g., {', '.join(model.species[:3])}"):
-    model.predict_species_within_audio_file(
+    next(model.predict_species_within_audio_file(
         TEST_FILE_WAV,
         filter_species=invalid_filter_species
-    )
+    ))
 
 
 def test_internal_predictions_are_correct(model: AudioModelV2M4Protobuf):