+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Separate `ModelV2M4Protobuf` into `AudioModelV2M4Protobuf` and `MetaModelV2M4Protobuf`
 - Separate `ModelV2M4` into `AudioModelV2M4` and `MetaModelV2M4`
 - Move v2.4 models to `birdnet.models.v2m4`
+- Yield results of `predict_species_within_audio_file` instead of returning an OrderedDict
 
 ## [0.1.5] - 2024-08-16